# domain-banking-cards-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
  /api/v1/cards/{cardId}/security/{securitySettingId}:
    put:
      tags:
      - Card Security
      summary: Update security settings
      description: Updates security settings for a card
      operationId: updateSecuritySettings
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: securitySettingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSecuritySettingsCommand"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/cards/{cardId}/limits/{limitId}:
    put:
      tags:
      - Card Limits
      summary: Update card limits
      description: Updates spending or withdrawal limits for a card
      operationId: updateLimits
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: limitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardLimitsCommand"
        required: true
      responses:
        "200":
          description: OK
...
  /api/v1/backoffice/cards/{cardId}:
    get:
      tags:
      - Card Backoffice
      summary: Get card details
      description: Retrieves full card information for backoffice review
      operationId: getCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
...
components:
  schemas:
    UpdateSecuritySettingsCommand:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        securitySettingId:
          type: string
          format: uuid
        securityFeature:
          type: string
        enabled:
          type: boolean
    UpdateCardLimitsCommand:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        limitId:
          type: string
          format: uuid
        limitType:
          type: string
        dailyLimit:
          type: number
        monthlyLimit:
          type: number
        transactionLimit:
          type: number
...
    CardDTO:
      required:
      - accountId
      - binId
      - cardHolderId
      - cardHolderName
      - cardNetworkId
      - cardNumber
      - cardStatus
      - cardTypeId
      - currencyCode
      - expirationMonth
      - expirationYear
      - isActive
      - isLocked
      - isPhysical
      - isPrimary
      - isVirtual
      - issuerId
      - partyId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        cardId:
          type: string
          format: uuid
        cardNumber:
          type: string
        maskedCardNumber:
          type: string
        cardSequenceNumber:
          type: string
        binId:
          type: string
          format: uuid
        cardTypeId:
          type: string
          format: uuid
        cardNetworkId:
          type: string
          format: uuid
        issuerId:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        partyId:
          type: string
          format: uuid
        cardStatus:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - CANCELLED
        cardHolderName:
          type: string
        cardHolderId:
          type: string
        expirationMonth:
          type: integer
          format: int32
        expirationYear:
          type: integer
          format: int32
        cvv:
          type: string
        pin:
          type: string
        activationDate:
          type: string
          format: date-time
        issuanceDate:
          type: string
          format: date-time
        expirationDate:
          type: string
          format: date-time
        lastUsedDate:
          type: string
          format: date-time
        isPhysical:
          type: boolean
        isVirtual:
          type: boolean
        isPrimary:
          type: boolean
        isActive:
          type: boolean
        isLocked:
          type: boolean
        lockReason:
          type: string
        dailyLimit:
          type: number
          format: double
        monthlyLimit:
          type: number
          format: double
        creditLimit:
          type: number
          format: double
        availableBalance:
          type: number
          format: double
        currencyCode:
          type: string
        designId:
          type: string
          format: uuid
        notes:
          type: string
...